[New plugin creator component] Filter by age
